🔥 [FEAT] modelo de reservacion creado y testeado
